selected: EB, unselected: REST
