+++ README.md
@@ -1,0 +1,16 @@
+### Selected Architecture: 
+
+EventBus - one constant connection 
+
+
+Event bus has lower latency which means a better user experience. Features are easier to add because the event bus structure decouples the features. Scalability is easier with event bus 
+
+
+### Unselected Architecture: 
+
+RESTFUL API - small procedure calls using api gateway. Clunkier than EventBus
+
+RESTFUL architecture has low latency and it is difficult to maintain real time features 
+
+
+
