+++ selected/bin/App.ts
@@ -4,7 +4,7 @@ import * as cdk from 'aws-cdk-lib';
 import { App } from '../lib/App';
 
 const app = new cdk.App();
-new App(app, 'App', {
+new App(app, 'eventbus', {
   /* If you don't specify 'env', this stack will be environment-agnostic.
    * Account/Region-dependent features and context lookups will not work,
    * but a single synthesized template can be deployed anywhere. */
