+++ selected/lambda/event-publisher/index.js
