+++ selected/lambda/game-processor/index.js
