+++ selected/lambda/websocket-handlers/connect.js
