+++ unselected/lambda/delete-game/index.js
@@ -1,64 +1,42 @@
-const AWS = require('aws-sdk');
-const dynamoDB = new AWS.DynamoDB.DocumentClient();
-const eventBridge = new AWS.EventBridge();
+import { DynamoDB } from '@aws-sdk/client-dynamodb';
+import { DynamoDBDocument } from '@aws-sdk/lib-dynamodb';
 
-exports.handler = async (event) => {
+const ddbClient = new DynamoDB({});
+const ddbDocClient = DynamoDBDocument.from(ddbClient);
+
+export const handler = async (event) => {
+  if (event.requestContext.http.method === 'OPTIONS') {
+    return { statusCode: 204, headers: corsHeaders() };
+  }
+
+  const gameId = event.pathParameters?.gameId;
+  if (!gameId) {
+    return {
+      statusCode: 400,
+      headers: corsHeaders(),
+      body: JSON.stringify({ message: 'Game ID is required' })
+    };
+  }
+
   try {
-    console.log('Event received:', JSON.stringify(event, null, 2));
-    
-    // Extract details from the event
-    const detail = JSON.parse(event.detail);
-    const connectionId = detail.connectionId;
-    const gameId = detail.gameId;
-    
-    if (!gameId) {
-      return {
-        statusCode: 400,
-        body: JSON.stringify({ message: 'Game ID is required' })
-      };
-    }
-    
-    // Delete the game from DynamoDB
-    await dynamoDB.delete({
+    await ddbDocClient.delete({
       TableName: process.env.DYNAMODB_TABLE,
       Key: {
         pk: `GAME#${gameId}`,
         sk: 'METADATA'
       }
-    }).promise();
-    
-    console.log(`Game ${gameId} deleted successfully`);
-    
-    // Publish an event to notify about the game deletion
-    await eventBridge.putEvents({
-      Entries: [{
-        Source: 'game.service',
-        DetailType: 'GameDeleted',
-        Detail: JSON.stringify({
-          gameId: gameId,
-          connectionId: connectionId,
-          deletedAt: new Date().toISOString()
-        }),
-        EventBusName: process.env.EVENT_BUS_NAME
-      }]
-    }).promise();
-    
+    });
+
     return {
       statusCode: 200,
-      body: JSON.stringify({ 
-        message: 'Game deleted successfully',
-        gameId: gameId
-      })
+      headers: corsHeaders(),
+      body: JSON.stringify({ message: 'Game deleted successfully' })
     };
   } catch (error) {
-    console.error('Error deleting game:', error);
-    
     return {
       statusCode: 500,
-      body: JSON.stringify({ 
-        message: 'Error deleting game',
-        error: error.message
-      })
+      headers: corsHeaders(),
+      body: JSON.stringify({ message: 'Error deleting game', error: error.message })
     };
   }
 };
