+++ unselected/package.json
@@ -34,7 +34,6 @@
     "constructs": "^10.4.2",
     "ionicons": "^7.4.0",
     "openid-client": "^5.7.0",
-    "source-map-support": "^0.5.21",
-    "uuid": "^9.0.1"
+    "source-map-support": "^0.5.21"
   }
 }
